# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "血癌照護管理平台",
        "description": "\n<h3>Application Interface Documentation for Blood Cancer Care Management Platform.</h3>\n<h4><a href='/'>Home Page</a></h4>\n                        ",
        "contact": {
            "name": "High Performance Distributed Systems Lab",
            "email": "nkust.ee.hpds@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/blood-components": {
            "post": {
                "tags": [
                    "抽血數據"
                ],
                "summary": "新增抽血數據",
                "description": "新增抽血數據",
                "operationId": "ea64f5c8eb7e998477cdc2f21aeff61f",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "account"
                                        ],
                                        "properties": {
                                            "account": {
                                                "description": "個案帳號",
                                                "type": "string",
                                                "example": "user1"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/blood"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "抽血數據編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/blood"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/blood-components/{id}": {
            "delete": {
                "tags": [
                    "抽血數據"
                ],
                "summary": "刪除抽血數據",
                "description": "刪除抽血數據",
                "operationId": "7416f250660c61628a7f6835f78a1f36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "抽血數據編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "patch": {
                "tags": [
                    "抽血數據"
                ],
                "summary": "更新抽血數據",
                "description": "更新抽血數據",
                "operationId": "0079f0491c12a52320ca5a2ea90a04e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "抽血數據編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/blood"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "抽血數據編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/blood"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/blood-components/account/{account}": {
            "get": {
                "tags": [
                    "抽血數據"
                ],
                "summary": "取得抽血數據",
                "description": "取得抽血數據",
                "operationId": "8ee664b6aba87ca2ea6021151282a403",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "個案帳號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "抽血數據編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/blood"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cases/{account}": {
            "get": {
                "tags": [
                    "個案資料"
                ],
                "summary": "取得個案資料",
                "description": "取得個案資料",
                "operationId": "ee988df769515cc5e3b8ee6e7f983417",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "個案帳號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "個案編號",
                                            "type": "integer"
                                        },
                                        "account": {
                                            "description": "個案帳號",
                                            "type": "string"
                                        },
                                        "transplant_num": {
                                            "description": "個案移植編號",
                                            "type": "string"
                                        },
                                        "name": {
                                            "description": "個案姓名",
                                            "type": "string"
                                        },
                                        "gender": {
                                            "description": "個案性別",
                                            "type": "string"
                                        },
                                        "birthday": {
                                            "description": "個案生日",
                                            "type": "string"
                                        },
                                        "date": {
                                            "description": "個案移植日期",
                                            "type": "string"
                                        },
                                        "transplant_type": {
                                            "description": "個案移植種類",
                                            "type": "string"
                                        },
                                        "disease_type": {
                                            "description": "個案疾病種類",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "blood": {
                "required": [
                    "wbc",
                    "hb",
                    "plt",
                    "got",
                    "gpt",
                    "cea",
                    "ca153",
                    "bun"
                ],
                "properties": {
                    "wbc": {
                        "description": "白血球(WBC)",
                        "type": "integer",
                        "example": "555"
                    },
                    "hb": {
                        "description": "血紅素(HB)",
                        "type": "integer",
                        "example": "555"
                    },
                    "plt": {
                        "description": "血小板(PLT)",
                        "type": "integer",
                        "example": "555"
                    },
                    "got": {
                        "description": "肝指數(GOT)",
                        "type": "integer",
                        "example": "555"
                    },
                    "gpt": {
                        "description": "肝指數(GPT)",
                        "type": "integer",
                        "example": "555"
                    },
                    "cea": {
                        "description": "癌指數(CEA)",
                        "type": "integer",
                        "example": "555"
                    },
                    "ca153": {
                        "description": "癌指數(CA153)",
                        "type": "integer",
                        "example": "555"
                    },
                    "bun": {
                        "description": "尿素氮(BUN)",
                        "type": "integer",
                        "example": "555"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "security": [
        {
            "basicAuth": []
        }
    ],
...
}